Crate DefaultConfigurablePropertys as a separate file
